+++ randovania/game_description/assignment.py
@@ -3,6 +3,7 @@ from typing import NamedTuple
 from randovania.game_description.requirements import Requirement
 from randovania.game_description.resources.pickup_entry import PickupEntry
 from randovania.game_description.resources.pickup_index import PickupIndex
+from randovania.game_description.world.dock import DockWeakness
 from randovania.game_description.world.node_identifier import NodeIdentifier
 
 
@@ -13,3 +14,4 @@ class PickupTarget(NamedTuple):
 
 PickupAssignment = dict[PickupIndex, PickupTarget]
 NodeConfigurationAssignment = dict[NodeIdentifier, Requirement]
+DockWeaknessAssignment = dict[NodeIdentifier, DockWeakness]