+++ randovania/games/prime1/gui/preset_settings/__init__.py
@@ -18,6 +18,7 @@ def prime1_preset_tabs(editor: PresetEditor, window_manager: WindowManager):
     from randovania.gui.preset_settings.location_pool_tab import PresetLocationPool
     from randovania.gui.preset_settings.metroid_item_pool_tab import MetroidPresetItemPool
     from randovania.games.prime1.gui.preset_settings.prime_generation_tab import PresetPrimeGeneration
+    from randovania.gui.preset_settings.dock_rando_tab import PresetDockRando
     return [
         PresetTrickLevel(editor, game_description, window_manager),
         PresetPatcherEnergy(editor, game_enum),
@@ -26,6 +27,7 @@ def prime1_preset_tabs(editor: PresetEditor, window_manager: WindowManager):
         PresetPrimeGeneration(editor, game_description),
         PresetPrimeGoal(editor),
         PresetPrimeHints(editor),
+        PresetDockRando(editor, game_description),
         PresetPrimePatches(editor),
         PresetLocationPool(editor, game_description),
         MetroidPresetItemPool(editor),