+++ randovania/generator/generator.py
@@ -1,17 +1,22 @@
 import asyncio
+import dataclasses
 from random import Random
-from typing import Optional, Callable, List, Dict
+from typing import Callable, Dict, List, Optional
 
 import tenacity
-
 from randovania.game_description import derived_nodes
-from randovania.game_description.assignment import PickupAssignment, PickupTarget
+from randovania.game_description.assignment import (PickupAssignment,
+                                                    PickupTarget)
 from randovania.game_description.game_description import GameDescription
 from randovania.game_description.game_patches import GamePatches
 from randovania.game_description.resources.pickup_entry import PickupEntry
 from randovania.game_description.world.world_list import WorldList
-from randovania.generator.filler.filler_library import filter_unassigned_pickup_nodes, UnableToGenerate
-from randovania.generator.filler.runner import run_filler, FillerPlayerResult, PlayerPool, FillerResults
+from randovania.generator import dock_weakness_distributor
+from randovania.generator.filler.filler_library import (
+    UnableToGenerate, filter_unassigned_pickup_nodes)
+from randovania.generator.filler.runner import (FillerPlayerResult,
+                                                FillerResults, PlayerPool,
+                                                run_filler)
 from randovania.generator.hint_distributor import PreFillParams
 from randovania.generator.item_pool import pool_creator
 from randovania.layout import filtered_database
@@ -21,7 +26,9 @@ from randovania.layout.generator_parameters import GeneratorParameters
 from randovania.layout.layout_description import LayoutDescription
 from randovania.layout.preset import Preset
 from randovania.resolver import resolver
-from randovania.resolver.exceptions import GenerationFailure, InvalidConfiguration, ImpossibleForSolver
+from randovania.resolver.exceptions import (GenerationFailure,
+                                            ImpossibleForSolver,
+                                            InvalidConfiguration)
 
 
 def _validate_item_pool_size(item_pool: List[PickupEntry], game: GameDescription,
@@ -45,6 +52,8 @@ async def create_player_pool(rng: Random, configuration: BaseConfiguration,
                                                                            num_players > 1,
                                                                            player_index=player_index,
                                                                            rng_required=rng_required)
+
+    base_patches = dock_weakness_distributor.distribute_pre_fill_weaknesses(base_patches)
 
     base_patches = await game_generator.hint_distributor.assign_pre_filler_hints(
         base_patches,
@@ -150,13 +159,13 @@ async def _create_pools_and_fill(rng: Random,
 
 
 def _distribute_remaining_items(rng: Random,
-                                filler_results: Dict[int, FillerPlayerResult],
-                                ) -> Dict[int, GamePatches]:
+                                filler_results: FillerResults,
+                                ) -> FillerResults:
     unassigned_pickup_nodes = []
     all_remaining_pickups = []
     assignments: Dict[int, PickupAssignment] = {}
 
-    for player, filler_result in filler_results.items():
+    for player, filler_result in filler_results.player_results.items():
         for pickup_node in filter_unassigned_pickup_nodes(filler_result.game.world_list.all_nodes,
                                                           filler_result.patches.pickup_assignment):
             unassigned_pickup_nodes.append((player, pickup_node))
@@ -178,10 +187,16 @@ def _distribute_remaining_items(rng: Random,
     for (node_player, node), (pickup_player, pickup) in zip(unassigned_pickup_nodes, all_remaining_pickups):
         assignments[node_player][node.pickup_index] = PickupTarget(pickup, pickup_player)
 
-    return {
-        index: filler_results[index].patches.assign_pickup_assignment(assignment)
-        for index, assignment in assignments.items()
-    }
+    return dataclasses.replace(
+        filler_results,
+        player_results={
+            player: dataclasses.replace(
+                result,
+                patches=result.patches.assign_pickup_assignment(assignments[player])
+            ) for player, result in filler_results.player_results.items()
+        }
+    )
+
     # FIXME: ignoring major-minor randomization
 
     # return {
@@ -217,10 +232,15 @@ async def _create_description(generator_params: GeneratorParameters,
 
     filler_results = await retrying(_create_pools_and_fill, rng, presets, status_update)
 
-    all_patches = _distribute_remaining_items(rng, filler_results.player_results)
+    filler_results = _distribute_remaining_items(rng, filler_results)
+    filler_results = await dock_weakness_distributor.distribute_post_fill_weaknesses(rng, filler_results, status_update)
+
     return LayoutDescription.create_new(
         generator_parameters=generator_params,
-        all_patches=all_patches,
+        all_patches={
+            player: result.patches
+            for player, result in filler_results.player_results.items()
+        },
         item_order=filler_results.action_log,
     )