+++ randovania/resolver/debug.py
@@ -103,6 +103,12 @@ def set_level(level: int):
 def debug_level() -> int:
     return _DEBUG_LEVEL
 
+@contextlib.contextmanager
+def with_level(level: int):
+    current_level = debug_level()
+    set_level(level)
+    yield
+    set_level(current_level)
 
 def debug_print(message: str):
     if _DEBUG_LEVEL > 0: