+++ randovania/resolver/logic.py
@@ -1,9 +1,10 @@
 from typing import Dict
 
 from randovania.game_description.game_description import GameDescription
-from randovania.game_description.requirements import RequirementSet
+from randovania.game_description.requirements import Requirement, RequirementSet
 from randovania.game_description.world.node import Node
 from randovania.layout.base.base_configuration import BaseConfiguration
+from randovania.resolver.state import State
 
 
 class Logic:
@@ -20,3 +21,7 @@ class Logic:
 
     def get_additional_requirements(self, node: Node) -> RequirementSet:
         return self.additional_requirements.get(node, RequirementSet.trivial())
+    
+    @property
+    def victory_condition(self) -> Requirement:
+        return self.game.victory_condition